+++ tests/rowCumprods.R
@@ -36,7 +36,7 @@ for (mode in c("logical", "integer", "double")) {
       if (setDimnames) dimnames(x) <- dimnames
       else dimnames(x) <- NULL    
       # Check names attribute
-      for (useNames in c(NA, TRUE, FALSE)) {
+      for (useNames in c(if (getRversion() < "4.4.0") NA, TRUE, FALSE)) {
         # Row/column ranges
         r0 <- rowCumprods_R(x, useNames = useNames)
         r1 <- rowCumprods(x, useNames = useNames)
@@ -64,7 +64,7 @@ for (mode in c("logical", "integer", "double")) {
     if (setDimnames) dimnames(x) <- dimnames
     else dimnames(x) <- NULL    
     # Check names attribute
-    for (useNames in c(NA, TRUE, FALSE)) {
+    for (useNames in c(if (getRversion() < "4.4.0") NA, TRUE, FALSE)) {
       # Row/column ranges
       r0 <- rowCumprods_R(x, useNames = useNames)
       r1 <- rowCumprods(x, useNames = useNames)
@@ -92,7 +92,7 @@ for (mode in c("logical", "integer", "double")) {
     if (setDimnames) dimnames(x) <- dimnames
     else dimnames(x) <- NULL    
     # Check names attribute
-    for (useNames in c(NA, TRUE, FALSE)) {
+    for (useNames in c(if (getRversion() < "4.4.0") NA, TRUE, FALSE)) {
       # Row/column ranges
       r0 <- rowCumprods_R(x, useNames = useNames)
       r1 <- rowCumprods(x, useNames = useNames)
@@ -122,7 +122,7 @@ for (mode in c("logical", "integer", "double")) {
     if (setDimnames) dimnames(x) <- dimnames
     else dimnames(x) <- NULL    
     # Check names attribute
-    for (useNames in c(NA, TRUE, FALSE)) {
+    for (useNames in c(if (getRversion() < "4.4.0") NA, TRUE, FALSE)) {
       # Row/column ranges
       r0 <- rowCumprods_R(x, useNames = useNames)
       r1 <- rowCumprods(x, useNames = useNames)
@@ -163,7 +163,7 @@ for (mode in c("logical", "integer", "double")) {
     if (setDimnames) colnames(x) <- colnames
     else dimnames(x) <- NULL
     # Check names attribute
-    for (useNames in c(NA, TRUE, FALSE)) {
+    for (useNames in c(if (getRversion() < "4.4.0") NA, TRUE, FALSE)) {
       r0 <- rowCumprods_R(x, useNames = useNames)
       r1 <- rowCumprods(x, useNames = useNames)
       r2 <- t(rowCumprods(t(x), useNames = useNames))
@@ -182,7 +182,7 @@ for (mode in c("logical", "integer", "double")) {
     if (setDimnames) rownames(x) <- rownames
     else dimnames(x) <- NULL
     # Check names attribute
-    for (useNames in c(NA, TRUE, FALSE)) {
+    for (useNames in c(if (getRversion() < "4.4.0") NA, TRUE, FALSE)) {
       r0 <- rowCumprods_R(x, useNames = useNames)
       r1 <- rowCumprods(x, useNames = useNames)
       r2 <- t(rowCumprods(t(x), useNames = useNames))