+++ tests/rowRanks.R
@@ -158,7 +158,7 @@ for (mode in c("integer", "double")){
     if (setDimnames) dimnames(x) <- dimnames
     else dimnames(x) <- NULL
     # Check names attribute
-    for (useNames in c(NA, TRUE, FALSE)) {
+    for (useNames in c(if (getRversion() < "4.4.0") NA, TRUE, FALSE)) {
       for (ties in c("max", "min", "average", "first", "last", "dense", "random")) {
         cat(sprintf("ties.method = %s\n", ties))
         # rowRanks():